ML-ResumeParsingBOT/requirements.txt
pdfplumber
python-docx
pandas
scikit-learn
streamlit
email-validator
python-dotenv
